requirements.txt
Django==4.2.1
mysqlclient==2.1.1
djangorestframework==3.14.0
gunicorn==20.1.0